pyproject.toml: Added pylint and black.
Pylint is a Python code checker.
Black is a Python code formatter.

Using either of these is non-committal at this stage; it is
planned that both will be tested upon 0.2.0 release, and 0.2.1
will be published shortly afterwards after both pylint and black
have had a good shot at the codebase.

0.2.1 will also publish a pylint/black usage guide.